+++ test/MultiSigWalletWithDailyLimit.js
@@ -56,7 +56,7 @@ contract('MultiSigWalletWithDailyLimit', (accounts) => {
             name: 'changeRequirement',
             type: 'function',
             inputs: [{
-                type: 'uint256',
+                type: 'uint8',
                 name: 'required'
             }]
         },
